feat: #526 inline babel-util into wrap-webpack-bundle
We will remove wrap-webpack-bundle in the future too.
