+++ packages/liferay-npm-build-support/src/loader/__tests__/__fixtures__/a-project/liferay-npm-bundler.config.js
