+++ packages/liferay-npm-build-support/src/loader/__tests__/__fixtures__/a-project/package.json
@@ -1,4 +1,4 @@
 {
-	"name": "babel-util",
+	"name": "a-project",
 	"version": "1.0.0"
 }
